updated the paymentOptionService tests
